Zu lange Kampfmeldung zerstört immer noch den Server.

+++ src/common/util/nrmessage.h
@@ -29,7 +29,7 @@ extern struct nrmessage_type * nrt_find(const struct locale *,
 extern const char * nrt_string(const struct nrmessage_type *type);
 extern const char * nrt_section(const struct nrmessage_type * mt);
 
-extern int nr_render(const struct message * msg, const struct locale * lang,
+extern size_t nr_render(const struct message * msg, const struct locale * lang,
                      char * buffer, size_t size, const void * userdata);
 extern int nr_level(const struct message *msg);
 extern const char * nr_section(const struct message *msg);
